Adds service/libpam when compiling open-vm-tools
For vSphere to recognize that vm-tools is running, vmtoolsd needs to be
started. Service files were added to start this service, based on how
the ubuntu package installs the service files:
https://salsa.debian.org/vmware-packaging-team/pkg-open-vm-tools
Also as VGAuthService is required by the service files, the libpam was
added to the compilation process, so this binary is created.
The process to install the compiled version of open-vm-tools will look
for the latest release in github and will download it, compile it and
install it. Using a concourse resource to get the latest release was not
feasible as the stemcell creation is a standalone process that does not
require external downloads.